Combined schema specs and custom specs
by searching for a custom spec in marathon.spec named by the same
field as the schema.  Also created some initial functions to
generate SRCs where some SRCs are more likely to be generated.
Thus, we have a greater chance of having stuff in scope between
demand and supply with a small number of generated values since
out of scope stuff is still nice to have in tests but less useful.
We'll probably want to shuffle the most likely SRCs, too.